boot: add netsetup.lua
This script sets the network up according to the wifi parameters, and
switches to SOFTAP mode in case of failure (after 10 seconds). In this
mode it sets the WiFi AP name to "iot4h-" followed by the chipid in
decimal, and the password is "password" followed by the flashid. It
serves DHCP by default so this allows to reconnect to the device over
telnet and reconfigure it for example.