svc: add telnet.lua
This is a basic telnet server similar to the one in most examples, it
injects via node.input() what it receives over the socket, and redirects
the console output using node.output() to a function which sends the
strings to the socket. It also prints the node's station address and
the prompt upon connect.