svc: telnet: implement local buffering
After upgrading the modules to version 1.5.4, the telnet server stopped
working. This issue is described in details in this comment :

  https://github.com/nodemcu/nodemcu-firmware/issues/730#issuecomment-154241161

In short the problem is that net.socket:send() doesn't implement EAGAIN
and used to buffer everything and to cause memory outages. Now it doesn't
do that anymore and it's up to the application to do it instead (which is
not much better). The issue above suggests that implementing a local buffer
and manipulating it via table.insert() and table.remove() is the most optimal
solution avoiding expensive memory copies. This is what this commit does.
Please note that table.insert() is misnamed, as it appends in reality.