svc: telnet: implement password-based protection
Now the first connection is responsible for authenticating. A password
is asked for. As long as it fails, the password is asked for again. Once
the password is valid, the shell is granted, and new connections can
enter in parallel without being asked anything.

The first line indicates the chipid, the IP address, and if the service is
currently authenticated or anonymous. Until authenticated, extra connections
are simply rejected.

The telnet password is found in variable "telnet_passwd" which should be
set in sysconf.lua. If it is not found, then the password defaults to the
concatenation of the chipid, the flashid and the flashsize, all verbatim
as reported by node.*() and printed in decimal without any delimiter. At
least it minimally protects against random connections. The port and
password are printed on the serial port at boot time.