svc: add a "recovery" function to detect if a recovery button is pressed
This will be useful to disable non-critical services and applications to
ease recovery. It is recommended to upload it as a service, preferably an
early one, such as "svc/02-recovery.lc".

It will check if the board defines a button named "brd_btn1", in this
case it will check if this button is pressed (gpio returning zero while
defaulting to a pull-up returning 1 when not pressed). It will then
return 1 if the button is pressed, otherwise zero. The init code will
check this before loading the main application code, and will interrupt
the boot process if the button is pressed.