boot: netsetup: call sntp_sync() once the WiFi is ready
This programs a time synchronization. It must not be done earlier due to
a bug in version 1.5.4.1 which causes a panic if attempted before WiFi is
up. It's possible that it will happen as well if the network goes down.
At least it's not called if the button is being pressed when the WiFi
gets up.