lib: add an hex file dump ("xxd")
This tool is able to hex dump an existing file (possibly a binary one).
This is convenient to download files or to inspect the validity of their
contents. The function is moderately big, it's better not to upload it
by default. The code is a bit complex because there appears to be no way
to print a single byte at once without a line feed, so we have to build
a string made of the concatenation of all bytes to display on a line,
without wasting time re-allocating objects for each concatenation.