boot: add preliminary support for LFS in init.lua
The LFS mechanism isn't trivial. It requires to upload a binary file
(output of luac.cross -f -o lfs.img) to the file system then have the
firmware reflash it to the LFS area. But this operation requires lots
of RAM and will usually fail if anything else was initialized. Hence
the need to do it first at boot when the file is present. Thus what
we do is that the file is checked first, and if found, is reflashed
and a reboot is operated. The file is renamed during operation so that
we don't enter an endless loop and we remove it after operations.

The second step is to actually call the LFS's _init() function to
load the embedded symbols. These appear under the "LFS." object and
can be enumerated using "LFS._list". There is one item per source
file, all having their own subfunctions as they appear in the original
files. In addition the files are seen as modules and are automatically
matched and returned by "require()" and "dofile()", though slashes are
not possible.

All global variables used as constants should be moved to LFS so that
they do not eat RAM space anymore. It's possible to see the list of
available ones this way:

    show(debug.getstrings())

Similarly the list of functions or variables from loaded modules in use
can be checked here:

    show(debug.getregistry())

These ones would also benefit from moving to LFS.