lib: rework the show() function
The show() function often reports errors when called on undumpable
types. Let's instead first make it dump the argument's value, then
try to iterate over its attributes and/or its metatable and dump
all of them. We now always sort the arguments so that it's not
needed anymore to keep both show and sshow. Thanks to this, userdata
are now dumpable. The object's type is also printed via tostring()
and strings are surrounded with double quotes even in table values,
removing any ambiguity when numbers appear in strings.