lib: clean up the xxd function
Using an array instead of building the string on the fly costs quite
some bytes in the function. Let's construct the string by concatenation
instead to shrink the function by ~25%. Local variables were placed
closer to their usage to save a few bytes, and the code was slightly
cleaned up to make it more readable.