lib: make the show() function work on any type of index
When tables mix string and non-string indexes, sort() fails due to the
lack of "<" operator. We cannot convert them to string either because
that would mean we'd lose the initial index and wouldn't be able to
print the value. One solution could consist in building an intermediary
table to map the string-based indexes to the original indexes, though
this remains costly for little benefit. Instead let's just protect the
table.sort() call so that in case of failure it stops the sort where it
failed and still returns what it could do. This is sufficient to achieve
this, which still lists all keys and values:

  x={["b"]=2,[1]="a",["a"]=1,[2]="b",[3]="c",["c"]=3}
  show(x)
  table: 0x3fff1640
   a                    1
   2                    "b"
   c                    3
   b                    2
   1                    "a"
   3                    "c"