lfs: add the _init and dummy_strings from the nodemcu firmware
These ones allow to build the global LFS table out of the SPIFFS that
was written to flash before rebooting. All elements of LFS are accessible
as LFS.<name>() where <name> is the name of original modules. Sadly these
ones are not listed in the LFS variable itself, and need to be enumerated
using LFS._list().