lfs: enumerate all available functions in LFS
This creates a symbol table made of all functions present in the LFS
table and adds it to the LFS index. This way it's possible to enumerate
all LFS functions using "show(LFS)". Doing so only consumes a few tens
of bytes and adds a lot of comfort so it makes sense to invest a bit of
RAM there. It may possibly speed up things by avoiding to call
node.flashindex() for each function lookup, though this wasn't measured.