boot: split init.lua into "init" and "start"
With LFS it becomes more convenient to think in terms of initialization
step and of service loader. This also allows to further shrink the init
code to the minimum needed to flash+upgrade a new LFS image, or load the
existing one.

"start" is the service startup sequence. It continues to look up certain
files and to load certain directories. It also supports this with LFS,
though files should be prefixed with "lib_" or "svc_" instead of "lib/"
and "svc/" in order to be automatically loaded.

Since there is backwards compatibility with previous init code, the old
init was replaced with the new one. It will just silently ignore anything
LFS-related if LFS is not supported.