svc: upgrade the oled service to support u8g2 and LFS
u8g2 is the only supported version in version 3.0 and is way better
than u8g in that it gets rid of the multi-pass mechanism so that it's
finally possible to use a regular frame buffer with much faster
operations.

The code was slightly changed to allow the driver's name to be configured
as a global variable, and to automatically detect the first available
font if not specified.

LFS support was added as well with a few functions exported. More than
30 kB of RAM are still available with the display running now.