boot: make the start code consistently print the loaded file name
There may be confusion when a file exists both as .lc and .lua. Also
when not using LFS, the extension was not printed in load_dir(), which
further adds to the confusion. Last, "appli.lua" was printed before
calling load_file() and during with no extension.

Let's consistently print the exact requested file name now.